Update `README.md`
Removing the part about quines taking no input, as this was wrong. In
Ken's original paper "Reflections on Trusting Trust" he only describes
"... a source program that, when compiled and executed, will produce as
output an exact copy of its source". There is no mention of the programs
not taking any input.